feat(theme): allow to configure website primary color (#411)
* feat(theme): allow to configure website primary color

* refactor(theme): default primary color, keep it required in the graphql schema

* docs(theme): default primary color

* fix(theme): to always merge theme options with default options

* refactor(theme): use black as the default primary color
